리뷰 틀 정도는 만듦 //  길만들기 > 주행완료 > 사진미션 > 리뷰 >... 만 하면 될 듯!!... 화이팅...~~!!

+++ Forrest/Begin_TableViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-let table_color = UIColor(red: CGFloat(24/225.0), green: CGFloat(24/225.0), blue: CGFloat(24/225.0), alpha: 1.0)
+let table_color = UIColor(red: CGFloat(24/225.0), green: CGFloat(24/225.0), blue: CGFloat(24/225.0), alpha: CGFloat(1.0))
 
 class Begin_TableViewController: UITableViewController
 {
